# PERF: improve StringArray.isna